"Improve parsing service and add spring-boot actuator"
In this commit, several changes have been introduced for the parsing service.

It was noticed that while parsing coin technicals from the driver, the flow was not very readable. Thus, the coin technicals parsing method has been refactored to improve readability and maintainability. Also, 'WebDriver' initialization was refactored to use 'EdgeOptions' for more flexibility and control, which comes in handy for future modifications.

In addition, Spring Boot Actuator has been added to the project dependencies which provides several useful features to monitor and manage the application.

Moreover, the application properties file has been updated to set 'spring.main.lazy-initialization' to true. This means that all the Spring Beans will not be initialized at startup, and will be initialized only when they are first requested for, hence, reducing the startup time.