"Refactor parsing service and CoinTechnicals implementation"
This commit represents significant changes in the parsing service and the handling of CoinTechnicals.

The method that parsed coin technicals in the ParsingService class has been moved directly into the parseTechnicals method. This improves code readability and eliminates the need for the extra method.

The CoinTechnicals class has been expanded to include inline sleeping on object creation to prevent neutral data returns. The structure has also been converted from a record to a standard class to encapsulate the logic into accessor methods, enhancing code maintainability and flexibility.

Furthermore, the parseMatic method in RestController now returns a new inner class instance "TechnicalsRecord", created within CoinTechnicals to retain the advantages of records. These changes are driven by the need for cleaner and more maintainable code.