Update data validation method in ParsingService
Updated the `waitUntilDataAreValid` method in ParsingService to `waitUntilDataAreUpdated`. This change addresses the issue of the parser sometimes returning old data due to the analysis results loading after the body has loaded on the website. Now, the parser waits until their indexes aren't loaded, and then includes the sleep() after that to ensure the data is current.