Added support for Kafka listeners
Added properties in application.properties file for bootstrap Kafka servers. Created CoinsListener.java that configures Kafka consumer properties and provides a KafkaListenerContainerFactory bean. Established a CoinTechnicalsListener with a method to listen to 'coins-topic' on Kafka.

Modified pom.xml to include the parser as a dependency and changed the parent artifact. This allows us to parse technical records received on Kafka and enables easier management of module versioning and dependencies.

These changes provide core support to consume and parse messages from Kafka.