The class 'TradingviewTechnicalParserApplication' was shortened to 'ParserApplication' and 'CoinsListener' was renamed to 'TechnicalsConsumerConfig' to make the classes context more clearer and align with the simplified package structure.
Also, the exceptions thrown by 'parseAll()' method in 'RestController.java' were reviewed and 'ExecutionException' was removed as it was obsolete.

In 'ParsingService.java', the file path for 'pairnames' was modified relative to the project root directory.

In 'TechnicalsConsumerConfig' the class Serialized has been changed to Deserializer to correct the Kafka consumer configuration.

Other minor changes have been done to improve code readability and maintainability.