Refactor parser config and integrate Selenium standalone service
ThreadConfig.java was deleted and ExecutorService was moved to newly created ParserConfig.java. ParsingService.java no longer builds WebDriver instance but accepts it as a Spring bean instead. The WebDriver now supports the RemoteWebDriver functionality, which will attempt to connect to a Selenium server address specified in the properties file. If it fails, it falls back to ChromeDriver. Docker-compose.yaml file was updated to include a standalone Selenium service.